Telegram bot tests for: async Telegram API implementation, save and restore chat.

+++ _docker/compose/down.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+if ! docker info > /dev/null 2>&1; then
+  echo "🐳 This script uses docker, and it isn't running - please start docker and try again!"
+  exit 1
+fi
+
 docker-compose -p subvt_monitoring_services -f 07-docker-compose-subvt-monitoring.yml down
 docker-compose -p subvt_polkadot_services -f 06-docker-compose-subvt-polkadot-services.yml down
 docker-compose -p subvt_kusama_services -f 05-docker-compose-subvt-kusama-services.yml down
